Refactor NodeVariantSelectionStrategy to simple cases and common namespace
