+++ Neos.ContentRepository.BehavioralTests/Tests/Behavior/Features/06-NodeDisabling/08-CreateNodeAggregateWithNodeWithDisabledAncestor_WithDimensions.feature
@@ -40,95 +40,6 @@ Feature: Creation of nodes underneath disabled nodes
     And the graph projection is fully up to date
     And VisibilityConstraints are set to "frontend"
 
-  Scenario: Create a new node with parent disabled with strategy onlyGivenVariant
-    Given the command DisableNodeAggregate is executed with payload:
-      | Key                          | Value              |
-      | nodeAggregateIdentifier      | "the-great-nodini" |
-      | coveredDimensionSpacePoint   | {"language":"de"}  |
-      | nodeVariantSelectionStrategy | "onlyGivenVariant" |
-    When the following CreateNodeAggregateWithNode commands are executed:
-      | nodeAggregateIdentifier | nodeTypeName                            | parentNodeAggregateIdentifier | nodeName     |
-      | nodingers-cat           | Neos.ContentRepository.Testing:Document | the-great-nodini              | pet-document |
-
-    Then I expect the node aggregate "nodingers-cat" to exist
-    And I expect this node aggregate to disable dimension space points []
-
-    When I am in dimension space point {"language":"mul"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to node cs-identifier;nodingers-cat;{"language":"mul"}
-    And I expect this node to be a child of node cs-identifier;the-great-nodini;{"language":"mul"}
-
-    When I am in dimension space point {"language":"de"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to no node
-
-    When I am in dimension space point {"language":"gsw"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to node cs-identifier;nodingers-cat;{"language":"mul"}
-    And I expect this node to be a child of node cs-identifier;the-great-nodini;{"language":"mul"}
-
-    When I am in dimension space point {"language":"ltz"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to node cs-identifier;nodingers-cat;{"language":"mul"}
-    And I expect this node to be a child of node cs-identifier;the-great-nodini;{"language":"ltz"}
-
-    When I am in dimension space point {"language":"en"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to node cs-identifier;nodingers-cat;{"language":"mul"}
-    And I expect this node to be a child of node cs-identifier;the-great-nodini;{"language":"mul"}
-
-    And the command EnableNodeAggregate is executed with payload:
-      | Key                          | Value              |
-      | nodeAggregateIdentifier      | "the-great-nodini" |
-      | coveredDimensionSpacePoint   | {"language":"de"}  |
-      | nodeVariantSelectionStrategy | "onlyGivenVariant" |
-    And the graph projection is fully up to date
-
-    When I am in dimension space point {"language":"de"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to node cs-identifier;nodingers-cat;{"language":"mul"}
-    And I expect this node to be a child of node cs-identifier;the-great-nodini;{"language":"mul"}
-
-  Scenario: Create a new node with parent disabled with strategy virtualSpecializations
-    Given the command DisableNodeAggregate is executed with payload:
-      | Key                          | Value              |
-      | nodeAggregateIdentifier      | "the-great-nodini" |
-      | coveredDimensionSpacePoint   | {"language":"de"}  |
-      | nodeVariantSelectionStrategy | "virtualSpecializations" |
-    When the following CreateNodeAggregateWithNode commands are executed:
-      | nodeAggregateIdentifier | nodeTypeName                            | parentNodeAggregateIdentifier | nodeName     |
-      | nodingers-cat           | Neos.ContentRepository.Testing:Document | the-great-nodini              | pet-document |
-
-    Then I expect the node aggregate "nodingers-cat" to exist
-    And I expect this node aggregate to disable dimension space points []
-
-    When I am in dimension space point {"language":"mul"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to node cs-identifier;nodingers-cat;{"language":"mul"}
-    And I expect this node to be a child of node cs-identifier;the-great-nodini;{"language":"mul"}
-
-    When I am in dimension space point {"language":"de"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to no node
-
-    When I am in dimension space point {"language":"gsw"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to no node
-
-    When I am in dimension space point {"language":"ltz"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to node cs-identifier;nodingers-cat;{"language":"mul"}
-    And I expect this node to be a child of node cs-identifier;the-great-nodini;{"language":"ltz"}
-
-    When I am in dimension space point {"language":"en"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to node cs-identifier;nodingers-cat;{"language":"mul"}
-    And I expect this node to be a child of node cs-identifier;the-great-nodini;{"language":"mul"}
-
-    And the command EnableNodeAggregate is executed with payload:
-      | Key                          | Value              |
-      | nodeAggregateIdentifier      | "the-great-nodini" |
-      | coveredDimensionSpacePoint   | {"language":"de"}  |
-      | nodeVariantSelectionStrategy | "virtualSpecializations" |
-    And the graph projection is fully up to date
-
-    When I am in dimension space point {"language":"de"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to node cs-identifier;nodingers-cat;{"language":"mul"}
-    And I expect this node to be a child of node cs-identifier;the-great-nodini;{"language":"mul"}
-
-    When I am in dimension space point {"language":"gsw"}
-    And I expect node aggregate identifier "nodingers-cat" and node path "document/pet-document" to lead to node cs-identifier;nodingers-cat;{"language":"mul"}
-    And I expect this node to be a child of node cs-identifier;the-great-nodini;{"language":"mul"}
-
   Scenario: Create a new node with parent disabled with strategy allSpecializations
     Given the command DisableNodeAggregate is executed with payload:
       | Key                          | Value              |
